fix: fix a bug about built-in database
the database won't be overwritten after updating

+++ app/src/main/java/cn/umafan/lib/android/util/network/model/UpdateBean.kt
@@ -7,5 +7,5 @@ data class UpdateBean(
     val info: UpdateInfoBean,
     var show: Boolean = false,
     var initiative: Boolean = false,
-    var currentDb: Long = 0L
+    var currentDb: Int = 0
 )
